Organize tool implementations
