+++ src/ShellMuse.Core/Planning/Tools/WriteFileTool.cs
@@ -3,13 +3,15 @@ using System.Text.Json;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace ShellMuse.Core.Planning;
+namespace ShellMuse.Core.Planning.Tools;
 
-public class ReadFileTool : ITool
+public class WriteFileTool : ITool
 {
     public Task<string> RunAsync(JsonElement args, CancellationToken cancellationToken = default)
     {
         var path = args.GetProperty("path").GetString() ?? string.Empty;
-        return Task.FromResult(File.Exists(path) ? File.ReadAllText(path) : "");
+        var content = args.GetProperty("content").GetString() ?? string.Empty;
+        File.WriteAllText(path, content);
+        return Task.FromResult("written");
     }
 }
